Rewrite the install script
[Prehistory]

The profile script was no more usable due to a undocumented behavior
of "dconf resulting in whether overriden profiles or error messages.
The logic of the script worked previously, but the "dconf" or
"gsettings" API changed breaking the script (forks of "terminal.sexy"(1))
which's logic is used by almost all GNOME Terminal theme projects.

I've reported #789056 to the official GNOME bug tracker for the "dconf
component which got closed with the following reason:

> This is simply an error in your script: you're using dconf to read keys
that have never been written to. You need to use the gsettings tool
instead, which takes into account the schema so it can return the default
value in that case.

I tried to find the documentation where this behavior of "dconf" is
described, but it seems like this knowledge is only documented in the
code itself or internal developer documents.

[Description]

Since the script was never really optimized I decided to rewrite it to
fix issues like GH-12 and GH-3 and implement the following additional
features:

  * Profile Handling - the script is now able to detect available
  profiles and
    * clone the default profile if no specific profile has been
    specified - The script is able to get the UUID of the "default"
    profile and clone it
    * install the theme for a specific profile - the user can pass the
    name of the profile the theme should be installed to
    * handle already existing "Nord" profiles via version comparison -
    if the "Nord" profile already exists and the script version is less
    than the installed version the user must confirm whether to override
    the theme of abort the installation, otherwise the profile will be
      * updated if the script version is greater than the installed
        version
      * reinstalled if the installed version is equal to the script
        version
  * Log Level - the script now provides a option to allow the user to
    define the log level. Available log levels are
      * 0 - ERROR
      * 1 - WARNING
      * 2 - SUCCESS (default)
      * 3 - INFO
      * 4 - DEBUG
  * Dependency Management -  the script is able to validate all required
    dependencies to ensure the script can run on the executive
    environment which includes
    * dconf (dconf)
    * expr (coreutils)
    * gsettings (glib2)
    * uuidgen (util-linux)

References:
  (1) https://terminal.sexy
  (2) https://bugzilla.gnome.org/show_bug.cgi?id=789056

Fixes GH-12 and GH-3

GH-13